[feat](ABC-1168): Button Menu - Add stretch attribute
